added/updated dynamics build definition/script

+++ src/Handyman.Dynamics/build.ps1
@@ -1,21 +1,36 @@
+[CmdletBinding(PositionalBinding = $false)]
+param(
+    [string]$artifacts = $null,
+    [switch]$ci
+)
+
 $ErrorActionPreference = 'Stop'
 
 $this = $PSScriptRoot
 $root = "$this/.."
-$artifacts = "$this/.artifacts/"
 $configuration = "release"
 
-Import-Module -Force -Scope Local "$root/common.psm1"
+if (!$artifacts) {
+    $artifacts = "$this/.artifacts/"
+    Remove-Item -Recurse $artifacts -ErrorAction Ignore
+}
 
-Remove-Item -Recurse $artifacts -ErrorAction Ignore
+Import-Module -Force -Scope Local "$root/common.psm1"
 
 exec dotnet build -c $configuration `
     "$this/src/Handyman.Dynamics/Handyman.Dynamics.csproj"
 
-exec dotnet test --configuration $configuration `
-    "$this/test/Handyman.Dynamics.Tests/Handyman.Dynamics.Tests.csproj"
+[string[]] $testArgs=@()
 
-exec dotnet pack --no-restore --no-build --configuration $configuration -o $artifacts `
+if ($ci) {
+    $testArgs += "--logger", "trx"
+}
+
+exec dotnet test --configuration $configuration `
+    "$this/test/Handyman.Dynamics.Tests/Handyman.Dynamics.Tests.csproj" `
+    $testArgs
+
+exec dotnet pack --no-restore --no-build -c $configuration -o $artifacts --include-symbols "-p:SymbolPackageFormat=snupkg" `
     "$this/src/Handyman.Dynamics/Handyman.Dynamics.csproj"
     
 write-host -f green 'script completed'
